Moved analyzer GUI related code to 'gui' subdirectory
