+++ src/analyzer/gui/mainwindow.h
