+++ src/analyzer/gui/mainwindow.ui
